Fix GHC loadArchive clang++ error on macOS
GHC 8.6.4 introduced an issue where linking to `-libc++` will cause
GHC to resolve the library patch to `c++`, which in turn is a link to
clang++.

This results in the following error:

```
ghc: loadArchive: Neither an archive, nor a fat archive: `/nix/store/?-clang-wrapper-5.0.2/bin/clang++'

<no location info>: error:
    loadArchive "/nix/store/?-clang-wrapper-5.0.2/bin/clang++": failed
```

This commit applies the relevant patch to revert GHC's linking behavior
on darwin for the applicable versions of GHC (i.e. after 8.6.4).

# Conflicts:
#	pkgs/development/compilers/ghc/8.10.7.nix
#	pkgs/development/compilers/ghc/9.0.1.nix
#	pkgs/development/compilers/ghc/9.2.1.nix